Feature: exclude columns
dmt already supports the ability to compare only a subset of columns,
and this PR supports the inverse.

I require this because there's an array column in my data that dmt
struggles to mock, so I want to exclude it. Currently, I have to list
about 15 columns to include, which isn't ideal if someone adds a new
column and forgets to update the mock input/output.