Add a description column to the raw customers table
Our tests only ever exclude the description column, so the values don't
need to match between the raw and expected tables. But it means we don't
run into ambiguous column errors like "description does not exist" when
running our tests. We'll want to handle this error better in the future though!