Feature: exclude columns (#70)
* Feature: exclude columns

dmt already supports the ability to compare only a subset of columns,
and this PR supports the inverse.

I require this because there's an array column in my data that dmt
struggles to mock, so I want to exclude it. Currently, I have to list
about 15 columns to include, which isn't ideal if someone adds a new
column and forgets to update the mock input/output.

* Add a description column to the raw customers table

Our tests only ever exclude the description column, so the values don't
need to match between the raw and expected tables. But it means we don't
run into ambiguous column errors like "description does not exist" when
running our tests. We'll want to handle this error better in the future though!